requirements.txt
# NEXUS-ATMS — Next-Generation Urban Intelligence Platform
# Complete Dependencies

# RL Framework (GPU-accelerated)
stable-baselines3>=2.1.0
gymnasium>=0.29.0
# torch — install separately with CUDA:
#   pip install torch --index-url https://download.pytorch.org/whl/cu124

# SUMO Interface
traci>=1.18.0
sumolib>=1.18.0

# Data & Visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Computer Vision
ultralytics>=8.0.0
opencv-python>=4.8.0
mediapipe>=0.10.0

# Backend & API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0
jinja2>=3.1.0
aiofiles>=23.0.0

# NLP (Natural Language Commands)
spacy>=3.6.0

# Graph Algorithms (Emergency Corridor)
networkx>=3.1

# Reporting & Carbon Certificates
reportlab>=4.0.0
Pillow>=10.0.0

# Voice Broadcast
gTTS>=2.3.0

# Audio (Acoustic Detection)
librosa>=0.10.0

# Utilities
tqdm>=4.65.0
tensorboard>=2.14.0
rich>=13.0.0
pygame>=2.5.0

# ML (Anomaly Detection, Classification)
scikit-learn>=1.3.0
joblib>=1.3.0

# Optional: Hyperparameter tuning
optuna>=3.3.0
